time: convert to 32-bit only, remove 2099 limit
- epoch: uint64_t -> uint32_t
- internal arithmetic: int64_t -> int32_t (typedef i32)
- year range: 1970-2099 -> 1970-2038 (32-bit safe, fits 8051)
- API signature unchanged, only parameter types updated

+++ include/elib_ft_time.h
@@ -28,14 +28,15 @@ typedef struct {
 
 /**
  * @brief Convert calendar time to Unix epoch seconds
- * @param t Calendar time (year 1970-2099)
+ * @param t Calendar time (year 1970-2038)
  * @param epoch Output epoch seconds (from 1970-01-01 00:00:00 UTC)
  * @return ELIB_FT_OK on success, error code otherwise
  *
- * Supported year range: 1970-2099 (uint32_t overflow-safe).
+ * Supported year range: 1970-2038 (32-bit safe).
+ * epoch fits in uint32_t (max: 2106-02-07 06:28:15).
  * Does not account for leap seconds or timezone (always UTC).
  */
-int elib_ft_time_to_epoch(const elib_ft_time_t *t, uint64_t *epoch);
+int elib_ft_time_to_epoch(const elib_ft_time_t *t, uint32_t *epoch);
 
 /**
  * @brief Convert Unix epoch seconds to calendar time
@@ -43,17 +44,17 @@ int elib_ft_time_to_epoch(const elib_ft_time_t *t, uint64_t *epoch);
  * @param t Output calendar time
  * @return ELIB_FT_OK on success, error code otherwise
  *
- * Supported year range: 1970-2099.
+ * Supported year range: 1970-2038 (32-bit safe).
  * Does not account for leap seconds or timezone (always UTC).
  */
-int elib_ft_epoch_to_time(uint64_t epoch, elib_ft_time_t *t);
+int elib_ft_epoch_to_time(uint32_t epoch, elib_ft_time_t *t);
 
 /**
  * @brief Get day of week from epoch seconds
  * @param epoch Epoch seconds
  * @return 0=Sunday, 1=Monday, ..., 6=Saturday
  */
-uint8_t elib_ft_time_wday(uint64_t epoch);
+uint8_t elib_ft_time_wday(uint32_t epoch);
 
 #ifdef __cplusplus
 }
